break out mustkeepnew* functions to pruning_test_helpers.go
